Midnight_backend/requirements.txt
flask~=2.3.0
python-dotenv~=1.0.0
psycopg2-binary
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
SQLAlchemy~=2.0.20
